requirements.txt
google-cloud-storage>=2.10.0
google-cloud-bigquery>=3.11.0
google-cloud-kms>=2.16.0
lxml>=4.9.0
gunicorn>=21.2.0
python-dateutil>=2.8.2